Generate related words to a word in lyrics.
Added a new method that takes a lyrics archive in form of words
of songs (the words are not in the order as they were in the song,
for copyright reasons). Then it takes each word from each song
and checks what other words are present with that word in songs.
Afterwards it takes the first 20 words that are mostly related to
any given word. Finally it writes one object for each word into
the Cloud Storage as a plaintext: the name of the file is the word
and the contents are related words, each on its own line, in order
of the most related to the least related.